+++ server/src/api/build.rs
@@ -1,5 +1,5 @@
-use crate::{models::*, DbId};
 use crate::AppData;
+use crate::{models::*, DbId};
 use actix_web::{delete, get, post, put, web, Error, HttpRequest, HttpResponse};
 use anyhow::Result;
 extern crate serde_derive;
@@ -7,6 +7,7 @@ extern crate serde_qs as qs;
 use crate::utils;
 use crate::STORAGE_PATH;
 use crate::STORAGE_TYPE;
+use actix_identity::Identity;
 use async_std::path::Path;
 use async_std::{io, prelude::*};
 use async_tar::Archive;
@@ -33,7 +34,11 @@ pub async fn get_all(req: HttpRequest, data: web::Data<AppData>) -> Result<HttpR
 
 // todo optimisations
 #[post("/build")]
-pub async fn post(req: HttpRequest, mut body: web::Payload, app_data: web::Data<AppData>) -> Result<HttpResponse, Error> {
+pub async fn post(
+    req: HttpRequest,
+    mut body: web::Payload,
+    app_data: web::Data<AppData>,
+) -> Result<HttpResponse, Error> {
     utils::validate_api_key(&req)?;
     // read post data / file
     let tmp_dir = tempfile::TempDir::new()?;
@@ -133,27 +138,45 @@ pub async fn post(req: HttpRequest, mut body: web::Payload, app_data: web::Data<
 
 #[delete("/build")]
 // pub async fn get_builds(req: HttpRequest, json_data: web::Json<utils::Paginate>, data: web::Data<AppData>) -> Result<HttpResponse, Error>{
-pub async fn delete(post_data: web::Json<utils::IdType>, app_data: web::Data<AppData>) -> Result<HttpResponse, Error> {
-    let conn = app_data.pool.get().expect("couldn't get db connection from pool");
-    let response = web::block(move || DbBuild::delete(&conn, post_data.id))
-        .await
-        .map_err(|e| {
+pub async fn delete(
+    post_data: web::Json<utils::IdType>,
+    auth: Identity,
+    app_data: web::Data<AppData>,
+) -> Result<HttpResponse, Error> {
+    debug!("{:?}", auth.identity());
+    if let Some(_auth) = auth.identity() {
+        let conn = app_data.pool.get().expect("couldn't get db connection from pool");
+        let response = web::block(move || DbBuild::delete(&conn, post_data.id))
+            .await
+            .map_err(|e| {
+                debug!("{}", e);
+                HttpResponse::InternalServerError().finish()
+            })?;
+
+        Ok(HttpResponse::Ok().json(response))
+    } else {
+        Ok(HttpResponse::Unauthorized().finish())
+    }
+}
+
+#[delete("/build/{id}")]
+pub async fn delete_id(
+    web::Path(id): web::Path<i32>,
+    auth: Identity,
+    app_data: web::Data<AppData>,
+) -> Result<HttpResponse, Error> {
+    debug!("{:?}", auth.identity());
+    if let Some(_auth) = auth.identity() {
+        let conn = app_data.pool.get().expect("couldn't get db connection from pool");
+        let response = web::block(move || DbBuild::delete(&conn, id)).await.map_err(|e| {
             debug!("{}", e);
             HttpResponse::InternalServerError().finish()
         })?;
 
-    Ok(HttpResponse::Ok().json(response))
-}
-
-#[delete("/build/{id}")]
-pub async fn delete_id(web::Path(id): web::Path<i32>, app_data: web::Data<AppData>) -> Result<HttpResponse, Error> {
-    let conn = app_data.pool.get().expect("couldn't get db connection from pool");
-    let response = web::block(move || DbBuild::delete(&conn, id)).await.map_err(|e| {
-        debug!("{}", e);
-        HttpResponse::InternalServerError().finish()
-    })?;
-
-    Ok(HttpResponse::Ok().json(response))
+        Ok(HttpResponse::Ok().json(response))
+    } else {
+        Ok(HttpResponse::Unauthorized().finish())
+    }
 }
 
 #[derive(Deserialize)]
@@ -164,14 +187,23 @@ pub struct BuildActive {
 
 // #[put("/build")]
 #[put("/build/active")]
-pub async fn active(post_data: web::Json<BuildActive>, app_data: web::Data<AppData>) -> Result<HttpResponse, Error> {
-    let conn = app_data.pool.get().expect("couldn't get db connection from pool");
-    let response = web::block(move || DbBuild::active(&conn, post_data.id, post_data.active))
-        .await
-        .map_err(|e| {
-            debug!("{}", e);
-            HttpResponse::InternalServerError().finish()
-        })?;
+pub async fn active(
+    post_data: web::Json<BuildActive>,
+    auth: Identity,
+    app_data: web::Data<AppData>,
+) -> Result<HttpResponse, Error> {
+    debug!("{:?}", auth.identity());
+    if let Some(_auth) = auth.identity() {
+        let conn = app_data.pool.get().expect("couldn't get db connection from pool");
+        let response = web::block(move || DbBuild::active(&conn, post_data.id, post_data.active))
+            .await
+            .map_err(|e| {
+                debug!("{}", e);
+                HttpResponse::InternalServerError().finish()
+            })?;
 
-    Ok(HttpResponse::Ok().json(response))
+        Ok(HttpResponse::Ok().json(response))
+    } else {
+        Ok(HttpResponse::Unauthorized().finish())
+    }
 }
